% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{auto_render}
\alias{auto_render}
\title{Re-render a document as it changes}
\usage{
auto_render(input, ..., renderer = NULL)
}
\arguments{
\item{input}{Filename of rmarkdown file}

\item{...}{Arguments passed on to \code{\link[rmarkdown:render]{rmarkdown::render()}}}

\item{renderer}{A function to call to render the file. Defaults to
\code{\link[rmarkdown:render]{rmarkdown::render()}} for \code{.md} or \code{.rmd} files and \code{\link[tinytex:lualatex]{tinytex::lualatex()}}
for \code{.tex} file. Provide a function for any other function.

Use Ctrl + break (windows), Esc (mac gui) or Ctrl + C (command  line) to stop
the watcher.}
}
\description{
Re-render a document as it changes
}
