% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binscatter.r
\name{binscatter}
\alias{binscatter}
\title{Make a binscatter plot, allowing for formulas and fixed effects}
\usage{
binscatter(formula, data, x = NULL, ..., weights = NULL)
}
\arguments{
\item{formula}{A formula to estimate, of the form \code{y ~ x1 + x2 | fe1 + fe2}}

\item{data}{Data to use}

\item{x}{Independent variable of interest. May be a vector (as in \code{binsreg()})
or the name of a varaible in \code{data}. If omitted, the first x variable in
the formula is used.}

\item{...}{Other arguments passed to \code{binsreg}}

\item{weights}{Optional vector of weights. Interpretation is the same as \code{\link[=lm]{lm()}}.}
}
\description{
This is a convenience wrapper around \code{\link[lfe:felm]{lfe::felm()}} and \code{\link[binsreg:binsreg]{binsreg::binsreg()}},
because I want the ease of Stata's binscatter.
}
\details{
\emph{Important note:} The standard errors here will not account for the fixed
effects. This function is useful for plots, and is not recommended for inference.
}
\examples{
res <- binscatter(Sepal.Length ~ Sepal.Width + Petal.Length | Species, iris)
print(res$bins_plot)
res <- binscatter(Sepal.Length ~ Sepal.Width + Petal.Length | Species, iris, x="Petal.Length")
}
\seealso{
\code{\link[binsreg:binsreg]{binsreg::binsreg()}}, \code{\link[lfe:felm]{lfe::felm()}}
}
