% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binscatter.r
\name{capture_plot_output}
\alias{capture_plot_output}
\title{Capture graphical output}
\usage{
capture_plot_output(expr)
}
\arguments{
\item{expr}{An expression to evaluate without printing output}
}
\value{
Whatever expr returned

This function takes advantage of R's delayed evaluation. \code{expr} isn't
valuated until it's forced in this function.
The output is printed to a temporary png, which is then deleted.
}
\description{
Capture graphical output
}
\seealso{
\code{\link[utils:capture.output]{utils::capture.output()}}
}
