% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_packages.r
\name{cite_attached_packages}
\alias{cite_attached_packages}
\title{Cite currently attached pacakges}
\usage{
cite_attached_packages(outfile = NULL, include = "base")
}
\arguments{
\item{outfile}{Write results to a file. The default, NULL, doesn't write a file.}

\item{include}{Extra packages to include, beyond those that have been
attached (ones that you've called \code{library()} for.)
The default, "base", includes a citation for R itself.}
}
\value{
A list of citations, invisibly

Note that some packages provide multiple citations. This function will pick only the
first one they provided.
}
\description{
Cite currently attached pacakges
}
\examples{
cite_attached_packages()
}
