% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_packages.r
\name{cite_everything}
\alias{cite_everything}
\title{Cite currently loaded pacakges}
\usage{
cite_everything(outfile = NULL, include = "base")
}
\arguments{
\item{outfile}{Write results to a file. The default, NULL, doesn't write a file.}

\item{include}{Extra packages to include, beyond those that have been loaded.
The default, "base", includes a citation for R itself.}
}
\value{
A list of citations

Note that some packages provide multiple citations. This function will pick only the
first one they provided.
}
\description{
Cite currently loaded pacakges
}
