% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{clear_all}
\alias{clear_all}
\title{Clear everything possible}
\usage{
clear_all(to_clear = c("graphics", "objects", "packages"))
}
\arguments{
\item{to_clear}{Vector of things to clear.
"graphics" means close all graphics devices.
"objects" means remove all objects from global environment.
"packages" means unload all loaded namespaces
The default is to do all three.}
}
\description{
Clear everything possible
}
