% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{configure_tikzDevice}
\alias{configure_tikzDevice}
\title{Set some tikzDevice options}
\usage{
configure_tikzDevice(font = "Libertinus Serif", cache_dir = "~/.R")
}
\arguments{
\item{font}{Name of a font, passed to \code{\\setmainfont}}

\item{cache_dir}{Directory to store \code{tikzDevice_cache} file}
}
\value{
Nothing
}
\description{
Set some tikzDevice options
}
