% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dropbox.r
\name{dropbox_home}
\alias{dropbox_home}
\title{Find the user's Dropbox folder}
\usage{
dropbox_home(dropbox_path_override = Sys.getenv("DROPBOX_PATH"))
}
\arguments{
\item{dropbox_path_override}{Override looking for the Dropbox path with this
(defaults to envionment variable DROPBOX_PATH, which is not an official Dropbox setting)}
}
\value{
A vector of the user's dropbox folder(s)
}
\description{
Find the user's Dropbox folder
}
