% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness.r
\name{ensure_id_vars}
\alias{ensure_id_vars}
\title{Raise an error if the claimed variables don't uniquely identify rows.}
\usage{
ensure_id_vars(df, ...)
}
\arguments{
\item{df}{A dataframe to test}

\item{...}{Passed to \link{is_id}}
}
\value{
Original \code{df}
}
\description{
Calls \link{is_id} (with warnings as errors), then returns the original data if \link{is_id}
returns \code{TRUE}.
}
