% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.r
\name{explain_analyze}
\alias{explain_analyze}
\title{EXPLAIN and ANALYZE the current query}
\usage{
explain_analyze(x, ...)
}
\arguments{
\item{x}{SQL query to explan and analyze}

\item{...}{Unused, included for compatability with \code{\link[dplyr:explain]{dplyr::explain()}}}
}
\value{
x, invisibly

Note that this is much more expensive than \code{explain()}
}
\description{
EXPLAIN and ANALYZE the current query
}
