% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{felm_strict}
\alias{felm_strict}
\title{Make felm slightly easier to use for someone who forgets to check things}
\usage{
felm_strict(formula, data, ..., dates_as_factor = TRUE, strict = TRUE)
}
\arguments{
\item{formula}{A formula to estimate}

\item{data}{Data to use}

\item{...}{Passed to \code{\link[lfe:felm]{lfe::felm()}}}

\item{dates_as_factor}{Should dates be converted to factors before running felm? (Default true)}

\item{strict}{Should felm's warnings be treated as errors? Default true.}
}
\value{
The result from \code{felm()}
}
\description{
Make felm slightly easier to use for someone who forgets to check things
}
