% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_R_packages.r
\name{find_rcpp_deps}
\alias{find_rcpp_deps}
\title{Find installed packages that depend on Rcpp}
\usage{
find_rcpp_deps(repos = getOption("repos", default =
  "https://cran.rstudio.com"), verbose = TRUE)
}
\arguments{
\item{repos}{CRAN repo to use (defaults to Rstudio's)}

\item{verbose}{Narrate progress? (default TRUE)}
}
\value{
Vector of installed pacakges that depend on Rcpp and are available in \code{repos}

Borrowed from https://github.com/RcppCore/rcpp-logs/blob/master/scripts/showReverseRcppDepends.r
}
\description{
Find installed packages that depend on Rcpp
}
