% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.r
\name{force_names}
\alias{force_names}
\title{Really get the names of the object}
\usage{
force_names(x)
}
\arguments{
\item{x}{Object to get names of}
}
\value{
names of x, or NULL if x has no names
}
\description{
Really get the names of the object
}
\seealso{
\code{\link[=names]{names()}}

Works even for dplyr's lazy tables.
}
