% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.r
\name{force_nrow}
\alias{force_nrow}
\title{Really get the number of rows of the table}
\usage{
force_nrow(.tbl)
}
\arguments{
\item{.tbl}{Table to count rows of}
}
\value{
Number of rows in .tbl
}
\description{
Really get the number of rows of the table
}
\seealso{
\code{\link[=nrow]{nrow()}}

Works even for dplyr's lazy tables.
}
