% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{get_ext}
\alias{get_ext}
\title{Get file's extension.}
\usage{
get_ext(path)
}
\arguments{
\item{path}{The file or connection's name}
}
\value{
The file or connection's extension

In most cases, just apply \code{\link[=file_ext]{file_ext()}}, but include a special case for connections.
It's not required that the file exists.
}
\description{
Get file's extension.
}
