% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{get_os}
\alias{get_os}
\title{Get the current operating system}
\usage{
get_os()
}
\value{
"win", "mac", or "linux".
Only works for Windows, Mac, and Linux. (Non-Mac unix systems are reported as Linux.)
}
\description{
Get the current operating system
}
