% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.r
\name{hol_daylight_saving}
\alias{hol_daylight_saving}
\title{US Daylight Saving Time}
\usage{
hol_daylight_saving(years)
}
\arguments{
\item{years}{Vector of years}
}
\value{
The date daylight saving started or ended for those years, as a named vector of dates.
}
\description{
Important note: these dates only return values when they were marked
at a federal level. For example, Indiana has done weird things.
"War time"  during World Wars 1 and 2 is not included.
}
\details{
These functions require the \code{lubridate} package.
}
