% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.r
\name{holidays}
\alias{holidays}
\alias{hol_us_federal_holidays}
\alias{hol_new_years_day}
\alias{hol_martin_luther_king_day}
\alias{hol_george_washington_birthday}
\alias{presidents_day}
\alias{hol_memorial_day}
\alias{hol_independence_day}
\alias{hol_labor_day}
\alias{hol_columbus_day}
\alias{hol_veterans_day}
\alias{hol_thanksgiving_day}
\alias{hol_christmas_day}
\alias{hol_good_friday}
\alias{hol_inauguration_day}
\alias{hol_easter}
\title{US holidays}
\usage{
hol_us_federal_holidays(years)

hol_new_years_day(years)

hol_martin_luther_king_day(years)

hol_george_washington_birthday(years)

presidents_day(years)

hol_memorial_day(years)

hol_independence_day(years)

hol_labor_day(years)

hol_columbus_day(years)

hol_veterans_day(years)

hol_thanksgiving_day(years)

hol_christmas_day(years)

hol_good_friday(years)

hol_inauguration_day(years)

hol_easter(years)
}
\arguments{
\item{years}{Vector of years}
}
\value{
Holidays for those years, as a named vector of dates.
}
\description{
Available holiday functions are:
}
\details{
\itemize{
\item \code{hol_us_federal_holidays()}: All US federal holidays.
These are: New Years day, Martin Luther King Jr. day,
Presidents day (George Washington's birthday), Memorial day,  Independence
day, Labor day, Columbus day, Veterans day, Thanksgiving day, and Christmas
day.
\item \code{hol_new_years_day()}: New Years day
\item \code{hol_martin_luther_king_day()}: Martin Luther King Jr. day
\item \code{hol_inauguration_day()}: Presidential Inauguration day (every four years)
\item \code{hol_george_washington_birthday()}: George Washington's birthday
\item \code{hol_presidents_day()}: Presidents day; same as \code{hol_george_washington_birthday()}
\item \code{hol_good_friday()}: Good Friday (not a federal holiday, but some markets are closed)
\item \verb{hol_easter():} Easter (not a federal holiday)
\item \code{hol_memorial_day()}: Memorial day
\item \code{hol_independence_day()}: Independence day (Fourth of July)
\item \code{hol_labor_day()}: Labor day
\item \code{hol_columbus_day()}: Columbus day
\item \code{hol_veterans_day()}: Veterans day
\item \code{hol_thanksgiving_day()}: Thanksgiving day
\item \code{hol_christmas_day()}: Christmas day
}

Important note: these holidays only return values when they were celebrated
at a federal level. For example, Veterans day is only 1938 onward.

These functions require the \code{lubridate} package.

A lot of this code comes from the tis package, but returns dates
}
