% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{install_lazy}
\alias{install_lazy}
\title{Install packages if not already installed}
\usage{
install_lazy(pkg_list, verbose = TRUE)
}
\arguments{
\item{pkg_list}{Vector of packages to install}

\item{verbose}{Narrate progress?}
}
\description{
Install packages if not already installed
}
