% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{is.connection}
\alias{is.connection}
\title{Test if connection (to a file/website/zipfile etc)}
\usage{
is.connection(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical: is it a connection?
}
\description{
Test if connection (to a file/website/zipfile etc)
}
