% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness.r
\name{is_id}
\alias{is_id}
\title{Do the claimed variables identify rows?}
\usage{
is_id(df, ..., notifier = base::warning)
}
\arguments{
\item{df}{A dataframe to test}

\item{...}{Variable names, following \link[dplyr:select]{dplyr::select} rules}

\item{notifier}{A function to report conditions you wouldn't want in an ID variable
(Defaults to \link[base:warning]{base::warning}. Other reasonable options might be \link[base:stop]{base::stop} to
escalate issues or \link[base:force]{base::force} to not report them.)}
}
\value{
TRUE/FALSE for ID-ness
}
\description{
Just like Stata's isid. For normal tables, this runs faster if data.table is
installed.
}
