% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.r
\name{lapply_bind_rows}
\alias{lapply_bind_rows}
\title{DEPRECATED: Run lapply then bind_rows, with multicore}
\usage{
lapply_bind_rows(X, FUN, ..., lapply_id = NULL, mc.cores = NULL)
}
\arguments{
\item{X}{Argument to FUN to iterate over}

\item{FUN}{Function to apply}

\item{...}{Other arguments passed to \code{FUN}}

\item{lapply_id}{If not NULL, create a column with the name provided by lapply_id to
identify the \code{X} value used.}

\item{mc.cores}{Number of cores to use. (Default of NULL tries to guess; see \link{get_cores})}
}
\value{
The result, as if passed to \code{\link[=lapply]{lapply()}}
}
\description{
DEPRECATED: Run lapply then bind_rows, with multicore
}
\seealso{
\code{\link[purrr:map]{purrr::map_df()}}, \code{\link[dplyr:bind]{dplyr::bind_rows()}}
}
