% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.r
\name{lapply_parallel}
\alias{lapply_parallel}
\title{DEPRECATED: Run lapply on multiple cores with mclapply}
\usage{
lapply_parallel(X, FUN, ..., mc.cores = NULL)
}
\arguments{
\item{X}{Argument to FUN to iterate over}

\item{FUN}{Function to apply}

\item{...}{Other arguments passed to \code{FUN}}

\item{mc.cores}{Number of cores to use. (Default of NULL tries to guess.)}
}
\value{
The result, as if passed to \code{\link[=lapply]{lapply()}}

This is now just a thin wrapper around furrr::future_map. You shou
}
\description{
DEPRECATED: Run lapply on multiple cores with mclapply
}
