% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{make_better_names}
\alias{make_better_names}
\title{Make names nicer to work with}
\usage{
make_better_names(x)
}
\arguments{
\item{x}{A character vector of names}
}
\value{
A transformed version of those names
}
\description{
Make names nicer to work with
}
\examples{
make_better_names(c("Country", "GDP $M", "Coast.Length"))
#> [1] "country" "gdp_mn"  "coast_length"
# Note that the guarantee to make the names unique can lead to some surprises
# for example, "a_and_b" becomes "a_and_b_3" in this case:
make_better_names(c("a and b", "a-and-b", "a.and.b", "a_and_b"))
#> c("a_and_b", "a_and_b_1", "a_and_b_2", "a_and_b_3")
# Here's a way to have a bad time:
make_better_names(c("", "x", "X", "x_1"))
}
\seealso{
\code{\link[=make.names]{make.names()}} and \code{\link[tibble:tibble]{tibble::tibble()}}'s \code{.name_repair} argument

Resulting names are guaranteed to be unique, and will almost certainly be
syntactic.
}
