% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{make_monthly}
\alias{make_monthly}
\title{Coarsen a date to monthly resolution}
\usage{
make_monthly(x)
}
\arguments{
\item{x}{A date vector}
}
\value{
A date vector of the same length, with the day-of-month always 1
}
\description{
Coarsen a date to monthly resolution
}
\examples{
make_monthly(as.Date(c("2000-01-01", "2000-01-02")))
#> [1] "2000-01-01" "2000-01-01"
}
