% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelsugar.r
\name{mclapply_socket}
\alias{mclapply_socket}
\title{Define a sockets version of mclapply}
\usage{
mclapply_socket(
  X,
  FUN,
  ...,
  mc.preschedule = TRUE,
  mc.set.seed = TRUE,
  mc.silent = FALSE,
  mc.cores = 1L,
  mc.cleanup = TRUE,
  mc.allow.recursive = TRUE
)
}
\arguments{
\item{X}{A vector/list to iterate over}

\item{FUN}{A function to apply to \code{X}}

\item{...}{What you pass to mclapply}

\item{mc.preschedule}{Not used, provided for consistency with \link[parallel:mclapply]{parallel::mclapply}}

\item{mc.set.seed}{Not used, provided for consistency with \link[parallel:mclapply]{parallel::mclapply}}

\item{mc.silent}{Not used, provided for consistency with \link[parallel:mclapply]{parallel::mclapply}}

\item{mc.cores}{Number of cores (default 1)}

\item{mc.cleanup}{Not used, provided for consistency with \link[parallel:mclapply]{parallel::mclapply}}

\item{mc.allow.recursive}{Not used, provided for consistency with \link[parallel:mclapply]{parallel::mclapply}}
}
\value{
lapply-like list
}
\description{
An implementation of \code{\link[parallel]{mclapply}} using \code{parallel::parLapply}
}
\details{
Windows does not support forking. This makes it impossible to use mclapply on Windows
to farm out work to additional cores.

NOTE: This calls a load-balancing function, which doesn't play well with random number
generation. Everything I'm doing is deterministic, but be careful.
If you need to change it, switch out the \code{clusterApplyLB} for \code{clusterApply}.
}
