% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{memory_limit}
\alias{memory_limit}
\title{Get or set memory limits}
\usage{
memory_limit(size = NA)
}
\arguments{
\item{size}{numeric. Request a new limit, in MiB.}
}
\value{
A vector with the (new) limit, in MiB.
}
\description{
For Linux (or BSD), this function calls \code{\link[unix:rlimit]{unix::rlimit_as()}}
For Windows, this function calls \code{\link[utils:memory.size]{utils::memory.limit()}}
For Mac OS X, no limiting is available.
}
\seealso{
\link[base]{Memory-limits} for other limits.
}
