% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.r
\name{merge_stata}
\alias{merge_stata}
\title{Use dplyr to emulate Stata merges.}
\usage{
merge_stata(x, y, multi = c("1:1", "1:m", "m:1", "1:1 _n"), by = NULL,
  keep = "inner", keepusing = NULL, ...)
}
\arguments{
\item{x}{A dataframe to merge ('master' in Stata)}

\item{y}{A dataframe to merge ('using' in Stata)}

\item{multi}{What multiple matches are allowed? One of "1:1", "1:m", "m:1",
or "1:1 _n". Note that m:m matches aren't allowed because they're
ill-defined. You can use standard dplyr joins if you instead want the
cartesian product (equivalent to Stata's joinby).}

\item{by}{Variables to join by (as strings)}

\item{keep}{Observations to keep. Can be specified with Stata words
(1, 2, 3, "master", "using", "match") or SQL words
("left", "right", "full", "semi").}

\item{keepusing}{Variables from the \code{y} ('using') dataset to keep.}

\item{...}{Other arguments passed onto the dplyr *_join function.}
}
\description{
Use dplyr to emulate Stata merges.
}
