% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{month_to_quarter}
\alias{month_to_quarter}
\title{Get quarter from month}
\usage{
month_to_quarter(mon, make_factor = FALSE)
}
\arguments{
\item{mon}{Month number vector (1-12)}

\item{make_factor}{Make the result into a factor with appropriate levels? (default FALSE)}
}
\value{
Quarters of the months (integer or factor)
}
\description{
Get quarter from month
}
