% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{narrate}
\alias{narrate}
\title{DEPRECATED: Message to the user}
\usage{
narrate(..., verbose = TRUE)
}
\arguments{
\item{...}{Arguments to \code{message}}

\item{verbose}{Actually use \code{message}? (default TRUE)}
}
\value{
NULL
Just a thin wrapper around \code{\link[=message]{message()}}
}
\description{
DEPRECATED: Message to the user
}
