% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.r
\name{nrow}
\alias{nrow}
\title{Really get the number of rows of the table}
\usage{
nrow(x, force = FALSE)
}
\arguments{
\item{x}{Table to count rows of}

\item{force}{Force the count of a \code{tbl_lazy}? (not used otherwise)}
}
\value{
Number of rows in .tbl
}
\description{
Really get the number of rows of the table
}
\seealso{
\code{\link[=nrow]{nrow()}}

Works even for dplyr's lazy tables.
}
