% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.r
\name{pg_add_foreign_key}
\alias{pg_add_foreign_key}
\title{Add a foreign key}
\usage{
pg_add_foreign_key(con, table_name, column_name, reftable, refcolumn)
}
\arguments{
\item{con}{A DBI connection to a postgresql database}

\item{table_name}{The name of a table in the postgresql database}

\item{column_name}{The column to put the foreign key constraint on}

\item{reftable}{The name of the table providing the foreign key constraint}

\item{refcolumn}{The name of the column providing the foreign key constraint}
}
\value{
Nothing
}
\description{
Add a foreign key
}
