% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.r
\name{pg_add_index}
\alias{pg_add_index}
\title{Create an index}
\usage{
pg_add_index(
  con,
  table_name,
  indexed_col,
  unique_index = FALSE,
  drop_existing = FALSE
)
}
\arguments{
\item{con}{A DBI connection to a postgresql database}

\item{table_name}{The name of a table in the postgresql database}

\item{indexed_col}{Columns or an expression on columns to index}

\item{unique_index}{Force index to be unique?}

\item{drop_existing}{Drop index if it already exists?}
}
\value{
The name of the created index

Note, postgres is smart enough that you don't need to index a column that's
already unique, but if you want to ALTER TABLE to make a primary key, you
have to start with a unique index.
}
\description{
Create an index
}
