% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.r
\name{pg_add_primary_key}
\alias{pg_add_primary_key}
\title{Add a primary key}
\usage{
pg_add_primary_key(con, table_name, key_col)
}
\arguments{
\item{con}{A DBI connection to a postgresql database}

\item{table_name}{The name of a table in the postgresql database}

\item{key_col}{The column to use as the primary key}
}
\value{
Nothing
}
\description{
Add a primary key
}
