% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.r
\name{pg_vacuum}
\alias{pg_vacuum}
\title{Run the VACUUM command, and maybe ANALYZE too.}
\usage{
pg_vacuum(con, table_name = NULL, analyze = TRUE)
}
\arguments{
\item{con}{A DBI connection to a postgresql database}

\item{table_name}{The name of a table in the postgresql database (
(The default processes all tables in the database)}

\item{analyze}{Run ANALYZE too? (Default TRUE)}
}
\value{
Nothing
}
\description{
Run the VACUUM command, and maybe ANALYZE too.
}
