% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{read_data}
\alias{read_data}
\title{DEPRECATED: Read a dataset into memory, using extension to figure out file type}
\usage{
read_data(file, ..., guess_max = 1e+06)
}
\arguments{
\item{file}{Filename to read}

\item{...}{Further arguments passed to the reader function}

\item{guess_max}{Number of rows to scan to check types in readxl
(defaults to 1,000,000, which will be slower than readxl's default of 1000)}
}
\value{
The read data
}
\description{
DEPRECATED: Read a dataset into memory, using extension to figure out file type
}
