% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{read_dta}
\alias{read_dta}
\title{Read and reconstitute Stata file.}
\usage{
read_dta(...)
}
\arguments{
\item{...}{Arguments passed to \link[haven:read_dta]{haven::read_dta}}
}
\value{
The data, as read by \link[haven:read_dta]{haven::read_dta}, but with less attribute nonsense.
}
\description{
Read and reconstitute Stata file.
}
