% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{remove_ext}
\alias{remove_ext}
\title{Return the filename without extension.}
\usage{
remove_ext(path)
}
\arguments{
\item{path}{The file(s) or connection's name(s) (works with vectors)}
}
\value{
The file or connection's name without extension

In most cases, just apply \code{\link[tools:fileutils]{tools::file_path_sans_ext()}}, but include a special case for connections.
It's not required that the file exists.
}
\description{
Return the filename without extension.
}
