% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{rename_cols}
\alias{rename_cols}
\title{Rename columns in a table}
\usage{
rename_cols(.tbl, .vars, strict = TRUE)
}
\arguments{
\item{.tbl}{Table to rename}

\item{.vars}{A named character vector, where the names are the new column
names and the values are existing column names.}

\item{strict}{Should the function raise an error if existing column names
can't be found? (Default TRUE)}
}
\value{
The same .tbl, with some renamed columns

Note that this function is the same as \verb{colnames()<-} for in-memory
data.frames, but also works for remote tbls. It's similar to pandas'
\code{.rename} method.
}
\description{
Rename columns in a table
}
\examples{
cols_to_rename <- c(cyl2 = "cyl")
rename_cols(mtcars, cols_to_rename)
}
