% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducible_graphics.r
\name{save_plot}
\alias{save_plot}
\title{Save a ggplot plot}
\usage{
save_plot(
  plot,
  filename,
  scale_mult = 1,
  bg = "transparent",
  device = NULL,
  reproducible = NULL
)
}
\arguments{
\item{plot}{The plot created by \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}}

\item{filename}{The filename to save (save type depends on extension)}

\item{scale_mult}{A scale multiplier on the size. Defaults to 1; bigger
numbers use a larger canvas.}

\item{bg}{The background color, passed to the output device. The default
is "transparent". If set to "transparent", the plot will be modified to make
the \code{panel.background}, \code{plot.background}, \code{legend.background}, and
\code{legend.box.background} transparent as well. Set it to "white" to retain
the normal ggplot behavior.}

\item{device}{The device to use. Default depends on filename extension. Uses
cairo_pdf devices when available. Use "tex" or "tikz" to save with \code{\link[tikzDevice:tikz]{tikzDevice::tikz()}}.}

\item{reproducible}{Logical, default FALSE unless the \code{SOURCE_DATE_EPOCH}
environment variable or \code{SOURCE_DATE_EPOCH} R option is set. If reproducible,
\code{save_plot} will attempt to create a reproducible output, with creation date
depending on \code{SOURCE_DATE_EPOCH}.}
}
\value{
The plot (invisibly)
}
\description{
Save a ggplot plot
}
\seealso{
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} https://reproducible-builds.org/docs/source-date-epoch/

Note that creating reproducible outputs currently depends on the system
command \code{sed}, which isn't installed by default on Windows.
This implementation of reproducible outputs is incompatible with multi-file
outputs, like tikz.
}
