% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{stopif}
\alias{stopif}
\title{Stop if a condition is TRUE}
\usage{
stopif(...)
}
\arguments{
\item{...}{Conditions that will be evaluated}
}
\description{
Stop if a condition is TRUE
}
\seealso{
\code{\link[=stopifnot]{stopifnot()}}
}
