% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_R_packages.r
\name{update_packages}
\alias{update_packages}
\title{Update installed packages}
\usage{
update_packages(repos = getOption("repos"), checkBuilt = TRUE,
  verbose = TRUE)
}
\arguments{
\item{repos}{CRAN repo to use (defaults to Rstudio's)}

\item{checkBuilt}{If TRUE, a package built under an earlier major.minor version of R
(e.g., 3.4) is considered to be ‘old’. (default TRUE)}

\item{verbose}{Narrate progress? (default TRUE)}
}
\value{
Character vector of packages updated (possibly of length 0)

This version of update_packages goes the extra step of reinstalling all packages that
depend on Rcpp whenever Rcpp is updated.
}
\description{
Update installed packages
}
