% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{vec2string}
\alias{vec2string}
\title{Turn a vector into one long string, used to generate messages}
\usage{
vec2string(x, quoted = TRUE, add_and = TRUE)
}
\arguments{
\item{x}{Vector to convert}

\item{quoted}{Should the individual elements have quotes around them?}

\item{add_and}{Should we add an 'and' before the last element}
}
\value{
A string (length 1 character vector)
}
\description{
Turn a vector into one long string, used to generate messages
}
\examples{
vec2string(c(1,2,3))  # "'1', '2', and '3'"
}
