% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.r
\name{winsorize}
\alias{winsorize}
\title{Winsorize one vector}
\usage{
winsorize(x, trim = 0.01)
}
\arguments{
\item{x}{Vector to winsorize (numeric)}

\item{trim}{Fraction of data to trim at top and bottom}
}
\value{
x, but winsorized

(this version deals with point masses!)
}
\description{
Winsorize one vector
}
