% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.r
\name{get_cores}
\alias{get_cores}
\title{Get a reasonable number for mc.cores}
\usage{
get_cores(max_cores = 10L)
}
\arguments{
\item{max_cores}{Max number of cores this function will return}
}
\value{
An integer number of cores to use
}
\description{
Get a reasonable number for mc.cores
}
