% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r, R/update_R_packages.r
\name{narrate}
\alias{narrate}
\alias{narrate}
\title{Message to the user}
\usage{
narrate(..., verbose = TRUE)

narrate(..., verbose = TRUE)
}
\arguments{
\item{...}{Arguments to \code{message}}

\item{verbose}{Actually use \code{message}? (default TRUE)}

\item{...}{Arguments to \code{message}}

\item{verbose}{Actually use \code{message}? (default TRUE)}
}
\value{
NULL
Just a thin wrapper around \code{\link[=message]{message()}}

NULL
Just a thin wrapper around \code{\link[=message]{message()}}
}
\description{
Message to the user

Message to the user
}
