% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{truncate_bytes}
\alias{truncate_bytes}
\title{Truncate a string to be no longer than a specific number of bytes}
\usage{
truncate_bytes(x, max_len = Inf)
}
\arguments{
\item{x}{Vector of strings to truncate}

\item{max_len}{Max length (in bytes). Can be one number or have the same lenth as \code{x}}
}
\value{
x, truncated to \code{max_len} bytes each. The function won't chop
multi-byte characters in half, so the result might be shorter than \code{max_len}
}
\description{
Truncate a string to be no longer than a specific number of bytes
}
\note{
Requires package \code{stringi}. This implementation is really inefficient.
}
\examples{
truncate_bytes(c("ab", "cde"), c(1, 2))
latin1_str <- "fa\xE7ile"
Encoding(latin1_str) <- "latin1"
truncate_bytes(latin1_str, 3)
truncate_bytes("ὯaὯa", 2) # empty string
truncate_bytes("ὯaὯa", 6) # Ὧa only
}
\seealso{
\code{\link[stringi:stri_sub]{stringi::stri_sub()}} and \code{\link[base:substr]{base::substr()}}
}
